+++ views/css/style.css
@@ -1,5 +1,9 @@
 
-.update {
-    text-align: center;
-    width: 18rem;
+.backgound{
+    background: linear-gradient(rgba(25, 27, 27, 0.7), rgba(26, 29, 29, 0.7)),
+    url("../../public/images/webImage.jpg");
+    background-size: cover;
+    background-repeat: no-repeat;
+    background-position: center center;
+    height: 100%;
 }
